Extract RequestSubscriber middleware
to follow Single Responsibilty and improve testing.
This also allows to introduce more subscriber with the same interface in the future.
