+++ lib/influxdb-rails.rb
@@ -2,12 +2,14 @@ require "net/http"
 require "net/https"
 require "rubygems"
 require "socket"
+require "influxdb/rails/middleware/request_subscriber"
 require "influxdb/rails/version"
 require "influxdb/rails/logger"
 require "influxdb/rails/exception_presenter"
 require "influxdb/rails/configuration"
 require "influxdb/rails/backtrace"
 require "influxdb/rails/rack"
+require "influxdb/rails/timestamp_conversion"
 
 require "influxdb/rails/railtie" if defined?(Rails::Railtie)
 
@@ -17,6 +19,7 @@ module InfluxDB
   module Rails
     class << self
       include InfluxDB::Rails::Logger
+      include InfluxDB::Rails::TimestampConversion
 
       attr_writer :configuration
       attr_writer :client
@@ -84,55 +87,11 @@ module InfluxDB
       end
       alias transmit report_exception
 
-      def handle_action_controller_metrics(_name, start, finish, _id, payload)
-        tags = configuration.tags_middleware.call({
-          status:      payload[:status],
-          format:      payload[:format],
-          http_method: payload[:method],
-          method:      "#{payload[:controller]}##{payload[:action]}",
-          server:      Socket.gethostname,
-          app_name:    configuration.application_name,
-        }.reject { |_, value| value.nil? })
-
-        ts = convert_timestamp(finish.utc)
-
-        begin
-          {
-            configuration.series_name_for_controller_runtimes => ((finish - start) * 1000).ceil,
-            configuration.series_name_for_view_runtimes       => (payload[:view_runtime] || 0).ceil,
-            configuration.series_name_for_db_runtimes         => (payload[:db_runtime] || 0).ceil,
-          }.each do |series_name, value|
-            client.write_point series_name, values: { value: value }, tags: tags, timestamp: ts
-          end
-        rescue StandardError => e
-          log :error, "[InfluxDB::Rails] Unable to write points: #{e.message}"
-        end
-      end
-
       # rubocop:enable Metrics/MethodLength
       # rubocop:enable Metrics/AbcSize
 
-      TIMESTAMP_CONVERSIONS = {
-        "ns" => 1e9.to_r,
-        nil  => 1e9.to_r,
-        "u"  => 1e6.to_r,
-        "ms" => 1e3.to_r,
-        "s"  => 1.to_r,
-        "m"  => 1.to_r / 60,
-        "h"  => 1.to_r / 60 / 60,
-      }.freeze
-      private_constant :TIMESTAMP_CONVERSIONS
-
-      def convert_timestamp(time)
-        conv = TIMESTAMP_CONVERSIONS.fetch(configuration.time_precision) do
-          raise "Invalid time precision: #{configuration.time_precision}"
-        end
-
-        (time.to_r * conv).to_i
-      end
-
       def current_timestamp
-        convert_timestamp(Time.now.utc)
+        convert_timestamp(Time.now.utc, configuration.time_precision)
       end
 
       def ignorable_exception?(ex)
