+++ spec/unit/influxdb_rails_spec.rb
@@ -9,105 +9,6 @@ RSpec.describe InfluxDB::Rails do
     end
   end
 
-  describe ".handle_action_controller_metrics" do
-    let(:start)   { Time.at(1_517_567_368) }
-    let(:finish)  { Time.at(1_517_567_370) }
-    let(:payload) { { view_runtime: 2, db_runtime: 2, controller: "MyController", action: "show", method: "GET", format: "*/*", status: 200 } }
-    let(:data)    do
-      {
-        values:    {
-          value: 2
-        },
-        tags:      {
-          method:      "MyController#show",
-          status:      200,
-          format:      "*/*",
-          http_method: "GET",
-          server:      Socket.gethostname,
-          app_name:    "my-rails-app",
-        },
-        timestamp: 1_517_567_370_000
-      }
-    end
-
-    context "application_name is set" do
-      it "sends metrics with taggings and timestamps" do
-        expect_any_instance_of(InfluxDB::Client).to receive(:write_point).with("rails.controller", data.merge(values: { value: 2000 }))
-        expect_any_instance_of(InfluxDB::Client).to receive(:write_point).with("rails.view", data)
-        expect_any_instance_of(InfluxDB::Client).to receive(:write_point).with("rails.db", data)
-
-        described_class.handle_action_controller_metrics("unused", start, finish, "unused", payload)
-      end
-    end
-
-    context "application_name is nil" do
-      before do
-        allow(InfluxDB::Rails.configuration).to receive(:application_name).and_return(nil)
-      end
-
-      it "does not add the app_name tag to metrics" do
-        tags = {
-          method:      "MyController#show",
-          status:      200,
-          format:      "*/*",
-          http_method: "GET",
-          server:      Socket.gethostname,
-        }
-
-        expect_any_instance_of(InfluxDB::Client).to receive(:write_point).with("rails.controller", include(tags: tags))
-        expect_any_instance_of(InfluxDB::Client).to receive(:write_point).with("rails.view", include(tags: tags))
-        expect_any_instance_of(InfluxDB::Client).to receive(:write_point).with("rails.db", include(tags: tags))
-
-        described_class.handle_action_controller_metrics("unused", start, finish, "unused", payload)
-      end
-    end
-
-    context "when tags_middleware is overwritten" do
-      before do
-        allow(InfluxDB::Rails.configuration).to receive(:tags_middleware).and_return(tags_middleware)
-      end
-
-      let(:tags_middleware) { ->(tags) { tags.merge(static: "value") } }
-
-      it "processes tags throught the middleware" do
-        tags = data[:tags].merge(static: "value")
-
-        expect_any_instance_of(InfluxDB::Client).to receive(:write_point).with("rails.controller", include(tags: tags))
-        expect_any_instance_of(InfluxDB::Client).to receive(:write_point).with("rails.view", include(tags: tags))
-        expect_any_instance_of(InfluxDB::Client).to receive(:write_point).with("rails.db", include(tags: tags))
-
-        described_class.handle_action_controller_metrics("unused", start, finish, "unused", payload)
-      end
-    end
-  end
-
-  describe ".convert_timestamp" do
-    let(:sometime) { Time.parse("2017-12-11 16:20:29.111222333 UTC") }
-    let(:configuration) { double("Configuration") }
-    before { allow(InfluxDB::Rails).to receive(:configuration).and_return configuration }
-
-    {
-      "ns" => 1_513_009_229_111_222_333,
-      nil  => 1_513_009_229_111_222_333,
-      "u"  => 1_513_009_229_111_222,
-      "ms" => 1_513_009_229_111,
-      "s"  => 1_513_009_229,
-      "m"  => 25_216_820,
-      "h"  => 420_280,
-    }.each do |precision, converted_value|
-      it "should return the timestamp in nanoseconds when precision is #{precision.inspect}" do
-        allow(configuration).to receive(:time_precision).and_return(precision)
-        expect(InfluxDB::Rails.convert_timestamp(sometime)).to eq(converted_value)
-      end
-    end
-
-    it "should raise an excpetion when precision is unrecognized" do
-      allow(configuration).to receive(:time_precision).and_return("whatever")
-      expect { InfluxDB::Rails.convert_timestamp(sometime) }
-        .to raise_exception(/invalid time precision.*whatever/i)
-    end
-  end
-
   describe ".current_timestamp" do
     it "should return the current timestamp in the configured precision" do
       now = Time.parse("2017-12-11 16:20:29.111222333 UTC")
